App/server/requirements.txt
# Django
Django>=4.2,<7.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# MySQL database connector
mysqlclient>=2.2.0

# Alternative MySQL connector (if mysqlclient fails to install)
# PyMySQL>=1.1.0

# Development tools
python-dotenv>=1.0.0
